Use OMPL to generate heuristics
The calculation is run at every planning cycle.
It does seem to slow down the planner slightly.
The calculation time seems to be dominated by trying to find
the exact coordinates of the goal, so this increased time
should go away when the more accurate heuristic is guiding
the search to where an analytic path solution can be found.